pyproject.toml
[project]
name = "log-analyzer-agent"
version = "0.1.0"
description = "A LangGraph-based agent for analyzing logs"
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.20",
    "pydantic>=2.0.0",
    "langchain-google-genai>=0.0.5",
    "langchain-groq>=0.0.1",
    "langchain-community>=0.0.10",
    "aiohttp>=3.8.5",
    "tavily-python>=0.1.9",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.7.0",
    "psycopg[binary,pool]>=3.1.0",
    "asyncpg>=0.28.0",
    "langgraph-checkpoint-postgres>=0.0.10",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.6",
    "pyjwt>=2.8.0",
    "email-validator>=2.0.0",
    "alembic>=1.13.0",
    "sqlalchemy>=2.0.0",
    "sse-starlette>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
start = "main:run_server"
dev = "main:run_dev_server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/log_analyzer_agent"]

[tool.hatch.build.targets.sdist]
include = ["src/", "scripts/", "*.py", "*.md", "*.txt", "*.toml", "*.yml", "*.yaml"]
